Implemanted accuracy functions to evaluate the quality of our model
In loss.py you can find three methods for computing the quality of segmentation. Intersection over Union, Recall, and Accuracy.

In loss.ipynb, there is a notebook to play with to see how the methods works (even though it is pretty straightforward). This notebook can be deleted later on, as it is not necessary to use loss.py
